+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/next/FileSystemNextModule.kt
@@ -11,6 +11,7 @@ import expo.modules.kotlin.types.Either
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import java.io.File
+import java.io.FileNotFoundException
 import java.io.FileOutputStream
 import java.net.URI
 
@@ -95,6 +96,18 @@ class FileSystemNextModule : Module() {
 
       Property("path") { file ->
         file.asString()
+      }
+
+      Property("md5") { file ->
+        try {
+          file.md5
+        } catch (e: FileNotFoundException) {
+          null
+        }
+      }
+
+      Property("size") { file ->
+        file.size
       }
     }